Update DigiPotX9Cxxx.cpp
Varibale _currentValue was of type uint8_t. On decrementing, it can't go negative and was cycling to the upper limit (255) of uint8_t. 

_currentValue += (direction == DIGIPOT_UP ? 1 : -1); // before

This won't affect the operating behavior of this digital potentiometer but at hardware level, this device does not cycle value which is mentioned in its datasheet as well. So it better to follow that behavior at default especially if one is using _currentValue from DigiPot::get().